@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.0+14
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,24 +29,20 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
   dio: ^4.0.0
-  path_provider: ^2.0.1
+  path_provider: ^2.0.2
   # epub: ^2.1.0
-  file_picker: ^3.0.1
+  file_picker: ^3.0.3
   sqflite: ^2.0.0+3
-  url_launcher: ^6.0.4
-  epub_viewer: ^0.2.1
-  stacked: ^2.1.6
-  stacked_services: ^0.8.6
-  get_it: ^6.1.1
+  url_launcher: ^6.0.7
+  epub_viewer: ^0.2.2
+  stacked: ^2.1.9
+  stacked_services: ^0.8.8
+  get_it: ^7.1.3
   pdf_viewer_jk: ^2.1.0
   pdf_text: ^0.5.0
-  async: ^2.5.0 
-  archive: ^3.1.2
-  quiver: ^3.0.1
-  xml: ^5.1.1
   flutter_staggered_animations: ^1.0.0
   intl: ^0.17.0
-  syncfusion_flutter_pdfviewer: ^19.1.67-beta
+  syncfusion_flutter_pdfviewer: ^19.1.69-beta
   lottie: ^1.0.1
   flutter_spinkit: ^5.0.0
   epubx: ^3.0.0-dev.3
@@ -54,8 +50,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.0.3
-  stacked_generator: ^0.4.6
+  build_runner: ^2.0.5
+  stacked_generator: ^0.4.8
   flutter_launcher_icons: ^0.9.0
   flutter_native_splash: ^1.1.8+4
   
